Add methods to choose the correct session cookie
Sometimes it can happen that a browser has two cookies for with the same name. This might be the case where developers forget to enforce a `path` variable or the `domain` attribute or accidentally call `Set-Cookie` twice. In those cases, this library would previously return whatever cookie  `r.Cookie(name)` would return.

This patch introduces a new function called `GetExact` which allows one to pass a filter function which receives the `Session` attribute. That way, there is a way to search for a specific session if multiple session have been found.

This functional change is fully backwards compatible by introducing a new interface `StoreExact` which defines what a store needs to implement to conform to this new logic. Stores which do not have these new methods will fallback to the previous behavior, were the first cookie found will be used.

All default stores in this library already implement the `StoreExact` method with this patch.